shared/requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.32.1
